+++ src/lib/modules/poe.ninja/currencyView/currencySubModules/submodules/FragmentModule.ts
@@ -1,4 +1,4 @@
-import PoeNinja from "../../../../AbstractClass/PoeNinja";
+import PoeNinja from "../../../../../AbstractClass/PoeNinja";
 
 /**
  * Represents a module for handling fragments in the PoeNinja currency view.
@@ -9,7 +9,10 @@ class FragmentModule extends PoeNinja {
    * @param league The league for which the module is created.
    * @param typeName The type name of the module.
    */
-  constructor(protected league: string, protected typeName: string) {
+  constructor(
+    protected league: string,
+    protected typeName: string,
+  ) {
     const type: string = "Fragment";
     super(league, typeName, type);
   }