+++ src/lib/modules/poe.watch/allModules/subModules/ScarabModule.ts
@@ -1,10 +1,9 @@
-import PoeWatch from "../../../AbstractClass/PoeWatch";
+import PoeWatch from "../../../../AbstractClass/PoeWatch";
 
 /**
  * Subclass of PoeWatch for accessing scarab data.
  */
 class ScarabModule extends PoeWatch {
-
   /**
    * Creates a new instance of ScarabModule.
    * @param league - The league from which the data will be fetched.